Refactor: split components
