+++ components/layouts/Layout.tsx
@@ -1,22 +1,15 @@
-import { useRouter } from 'next/router';
 import React from 'react';
 import Navigation from './Navigation';
 import styled from '@emotion/styled';
 import Footer from './Footer';
+import { MEDIA_QUERY_END_POINT } from '../../constants';
+import Title from './components/Layout/Title';
 
 export const Layout = ({ children }: any) => {
-  const router = useRouter();
-  let title = '';
-  if (router.pathname === '/location') {
-    title = '매장안내';
-  } else {
-    title = '알라카르테';
-  }
-
   return (
     <Container>
       <Navigation />
-      <Title>{title}</Title>
+      <Title />
       <ContentContainer>{children}</ContentContainer>
       <Footer />
     </Container>
@@ -25,17 +18,19 @@ export const Layout = ({ children }: any) => {
 
 const Container = styled.div`
   min-width: 390px;
-`;
-
-const Title = styled.h1`
-  font-size: 1.8rem;
-  font-weight: bold;
-  text-align: center;
-  margin: 3.75rem 0;
+  button {
+    display: none;
+    @media screen and (max-width: ${MEDIA_QUERY_END_POINT.MOBILE}) {
+      display: block;
+    }
+  }
 `;
 
 const ContentContainer = styled.main`
   max-width: 1280px;
   margin: auto;
   padding: 0 2rem;
+  @media screen and (max-width: ${MEDIA_QUERY_END_POINT.MOBILE}) {
+    pading: 0 0.5rem;
+  }
 `;
